Separated the warm-up code from the actual code.
